Refactor MainWindow into mixins
